# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Arc Pass API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /eligibility/check:
    post:
      operationId: checkEligibility
      tags: [eligibility]
      summary: Check eligibility preview for Arc Pass
      description: >
        Public preview only. Does not authenticate the caller and does not
        grant claim access — identity verification via OAuth is required
        before anything can be claimed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EligibilityQuery"
      responses:
        "200":
          description: Eligibility preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EligibilityResult"
  /eligibility/preview:
    post:
      operationId: previewEligibility
      tags: [eligibility]
      summary: Get a privacy-safe public eligibility preview
      description: A username is not ownership proof. The response contains status values only and can never establish new Onchain Builder eligibility.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EligibilityQuery"
      responses:
        "200":
          description: Minimal eligibility preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EligibilityResult"
...
  /admin/auth/login:
    post:
      operationId: adminLogin
      tags: [admin-auth]
      summary: Start a dedicated, HTTP-only administrator session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminCredentials" }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSession" }
        "401":
          description: Invalid credentials or locked account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, username, displayName, avatarUrl, provider, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        provider:
          type: string
          enum: [x, discord]
        providerId:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    EligibilityQuery:
      type: object
      required: [identifier, platform]
      properties:
        identifier:
          type: string
          description: Username on the chosen platform (preview only, not an ownership proof)
        platform:
          type: string
          enum: [x, discord]
    EligibilityResult:
      type: object
      required: [founder, builder]
      properties:
        founder:
          type: object
          required: [status]
          properties:
            status:
              type: string
              enum: [eligible, ineligible, under_review, claimed, unknown]
        builder:
          type: object
          required: [status]
          properties:
            status:
              type: string
              enum: [verification_required, claimed, unknown]
...
    AdminCredentials:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string, format: password }
    AdminSession:
      type: object
      required: [id, email, role]
      properties:
        id: { type: integer }
        email: { type: string }
        role: { type: string }
...